:construction_worker: Reorganizing templates, using bolt.workspace and cleaning warning messages
