+++ cli/Cargo.toml
@@ -13,10 +13,6 @@ edition = { workspace = true }
 name = "bolt"
 path = "src/bin/main.rs"
 
-[profile.release]
-opt-level = 3
-lto = true
-
 [features]
 dev = []
