+++ cli/src/templates/program/constants.rs
@@ -1,0 +1,4 @@
+use anchor_lang::prelude::*;
+
+#[constant]
+pub const SEED: &str = "anchor";
