+++ cli/src/templates/program/error.rs
@@ -1,0 +1,7 @@
+use anchor_lang::prelude::*;
+
+#[error_code]
+pub enum ErrorCode {
+    #[msg("Custom error message")]
+    CustomError,
+}
