+++ cli/src/templates/program/instructions/initialize.rs
@@ -1,0 +1,8 @@
+use anchor_lang::prelude::*;
+
+#[derive(Accounts)]
+pub struct Initialize {}
+
+pub fn handler(_ctx: Context<Initialize>) -> Result<()> {
+    Ok(())
+}
