+++ cli/src/templates/program/instructions/mod.rs
@@ -1,0 +1,3 @@
+pub mod initialize;
+
+pub use initialize::*;
