+++ cli/src/templates/workspace/gitignore
@@ -1,0 +1,8 @@
+.anchor
+.bolt
+.DS_Store
+target
+**/*.rs.bk
+node_modules
+test-ledger
+.yarn
