+++ cli/src/templates/workspace/prettierignore
@@ -1,0 +1,8 @@
+.anchor
+.bolt
+.DS_Store
+target
+node_modules
+dist
+build
+test-ledger
