+++ cli/src/templates/workspace/workspace.toml
@@ -1,0 +1,28 @@
+[workspace]
+members = [
+    "programs/*",
+    "programs-ecs/components/*",
+    "programs-ecs/systems/*"
+]
+resolver = "2"
+
+[workspace.dependencies]
+bolt-lang = "{VERSION}"
+anchor-lang = "{ANCHOR_VERSION}"
+
+[profile.release]
+overflow-checks = true
+lto = "fat"
+codegen-units = 1
+
+[profile.release.build-override]
+opt-level = 3
+incremental = false
+codegen-units = 1
+
+[patch.crates-io]
+# Uncomment this to use the latest in-development version of Bolt
+# bolt-lang = {{ git = "https://github.com/magicblock-labs/bolt.git", branch = "main" }}
+
+# Uncomment this to use a local version of Bolt for development (bolt must be cloned above the current directory)
+# bolt-lang = {{ path = "../bolt/crates/bolt-lang"}}
