+++ programs/bolt-system/Cargo.toml
@@ -19,6 +19,9 @@ no-log-ix-name = []
 cpi = ["no-entrypoint"]
 default = []
 idl-build = ["anchor-lang/idl-build"]
+anchor-debug = ["anchor-lang/anchor-debug"]
+custom-heap = []
+custom-panic = []
 
 [dependencies]
 anchor-lang = { workspace = true }
